SBMLの変換モジュールを内製のものに切り替えました。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@2576 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/EcellLib/EcellLib/Project.cs
@@ -422,6 +422,9 @@ namespace Ecell
         {
             // Delete empty project.
             string prjPath = Path.Combine(Util.GetBaseDir(), m_info.Name);
+            if (!Directory.Exists(prjPath))
+                return;
+
             string[] files = Directory.GetFiles(prjPath, "project.xml");
             if (files.Length <= 0)
                 Directory.Delete(prjPath, true);
